refactor: CarNameParser 역할을 RacingCarGameRunner 에서 분리
