+++ kotlin-playground/src/main/kotlin/org/gonza/kotlinplayground/service/Round.kt
@@ -6,6 +6,7 @@ import org.gonza.kotlinplayground.domain.car.RacingCars
 import org.gonza.kotlinplayground.service.dto.GameResult
 import org.gonza.kotlinplayground.service.vo.CarName
 import org.gonza.kotlinplayground.service.vo.TryCount
+import org.gonza.kotlinplayground.util.CarNameParser
 
 class Round(
     private val carList: List<Car>,
@@ -33,27 +34,14 @@ class Round(
         )
     }
 
-    fun findWinner(): String {
+    fun findWinner(): List<CarName> {
         val racingCars = RacingCars(carList)
-        val winnerCarNameList = racingCars.findFarthestCarNameList()
-        return CarNameParser.toCarNameString(winnerCarNameList)
+        return racingCars.findFarthestCarNameList()
     }
 
     fun isFinished(maxTryCount: TryCount): Boolean = maxTryCount.graterThanOrEqual(currentTryCount)
 
     // FIXME: 여기 부분을 제거해야할 것 같음, Round 라는 도메인에 걸맞지 않음
-    private object CarNameParser {
-        private const val SPLIT_KEYWORD = ","
-
-        fun parse(carName: CarName): List<CarName> =
-            carName.value
-                .split(SPLIT_KEYWORD)
-                .map { it.trim() }
-                .map { CarName(it) }
-
-        fun toCarNameString(carNameList: List<CarName>): String = carNameList.joinToString(SPLIT_KEYWORD) { it.value }
-    }
-
     private object CarNameConverter {
         fun toCarList(carNameList: List<CarName>): List<Car> = carNameList.map { Car(it.value, 0) }
     }
