+++ kotlin-playground/src/main/kotlin/org/gonza/kotlinplayground/util/CarNameParser.kt
@@ -1,0 +1,15 @@
+package org.gonza.kotlinplayground.util
+
+import org.gonza.kotlinplayground.service.vo.CarName
+
+class CarNameParser {
+    companion object {
+        private const val SPLIT_KEYWORD = ","
+
+        fun parse(carName: CarName): List<CarName> =
+            carName.value
+                .split(SPLIT_KEYWORD)
+                .map { it.trim() }
+                .map { CarName(it) }
+    }
+}
